100% norm done.Created 2 new files check_1_1.c and check_1_2.c
